I added this branch per Gavin's request. This shows how to get DNK to work under medium trust. I had to tear enterprise library out and make some config changes to make this happen, so we thought it would be best in its own branch.

+++ Incremental.Kick/Helpers/PasswordGenerator.cs
@@ -2,9 +2,12 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Incremental.Kick.Helpers {
-    public class PasswordGenerator {
-        public static string Generate(byte passwordLength) {
+namespace Incremental.Kick.Helpers
+{
+    public class PasswordGenerator
+    {
+        public static string Generate(int passwordLength)
+        {
             char[] vowels = new char[] { 'a', 'e', 'i', 'o', 'u' };
             char[] consonants = new char[] { 'b', 'c', 'd', 'f', 'g', 'h', 'j', 'k', 'l', 'm', 'n', 'p', 'r', 's', 't', 'v' };
             char[] doubleConsonants = new char[] { 'c', 'd', 'f', 'g', 'l', 'm', 'n', 'p', 'r', 's', 't' };
@@ -13,25 +16,35 @@ namespace Incremental.Kick.Helpers {
 
             StringBuilder password = new StringBuilder();
 
-            for (int i = 0; i < passwordLength; i++) {
-                if ((password.Length > 0) && (!haveWrittenConsonant) && (ThreadSafeRandom.Next(100) < 10)) {
+            for (int i = 0; i < passwordLength; i++)
+            {
+                if ((password.Length > 0) && (!haveWrittenConsonant) && (ThreadSafeRandom.Next(100) < 10))
+                {
                     password.Append(doubleConsonants[ThreadSafeRandom.Next(doubleConsonants.Length)], 2);
 
                     haveWrittenConsonant = true;
-                } else {
-                    if ((!haveWrittenConsonant) && (ThreadSafeRandom.Next(100) < 90)) {
+                }
+                else
+                {
+                    if ((!haveWrittenConsonant) && (ThreadSafeRandom.Next(100) < 90))
+                    {
                         password.Append(consonants[ThreadSafeRandom.Next(consonants.Length)]);
                         haveWrittenConsonant = true;
-                    } else {
+                    }
+                    else
+                    {
                         password.Append(vowels[ThreadSafeRandom.Next(vowels.Length)]);
                         haveWrittenConsonant = false;
                     }
                 }
             }
 
-            if (password.Length > passwordLength) {
+            if (password.Length > passwordLength)
+            {
                 return password.ToString(0, passwordLength);
-            } else {
+            }
+            else
+            {
                 return password.ToString();
             }
         }
